refactor: Phase 3 - Extract player combat and movement systems
- Created src/systems/ module with CombatSystem and MovementSystem classes
- Refactored player.py from 915 lines to 405 lines (55% reduction)
- Extracted 279 lines of combat logic to src/systems/combat.py
- Extracted 361 lines of movement logic to src/systems/movement.py
- Maintained backward compatibility through property delegation
- All combat functionality (melee/ranged attacks, damage, projectiles) now in CombatSystem
- All movement functionality (pathfinding, mouse clicks, animation) now in MovementSystem
- Game tested and verified to work correctly with refactored architecture
- Follows composition over inheritance pattern for better maintainability

This completes Phase 3 of the refactoring plan as outlined in REFACTORING_PHASES.md
