+++ src/systems/__init__.py
@@ -1,0 +1,11 @@
+"""
+Systems module for the RPG game.
+
+This module contains various game systems that can be composed together
+to create complex game behaviors.
+"""
+
+from .combat import CombatSystem
+from .movement import MovementSystem
+
+__all__ = ['CombatSystem', 'MovementSystem']
